SHT-272: - fix "cannot losslessly convert 8.6 to int64"

+++ internal/admin/queries/orders.sql
@@ -7,7 +7,7 @@ SELECT o.id                       AS id,
        o.created_at               AS created_at,
        o.updated_at               AS updated_at,
        wa.currency                AS currency,
-       (SELECT SUM(p.price * op.quantity)
+       (SELECT SUM(p.price * op.quantity)::float8
         FROM order_products op
                  JOIN products p ON p.id = op.product_id
         WHERE op.order_id = o.id) AS total_price,
